# src/main/java/com/example/user/utils/model/Page.java
package com.example.user.utils.model;

import io.swagger.v3.oas.annotations.media.Schema;

public record Page(
        @Schema(description = "page number")
        long pageNumber,
        @Schema(description = "page size")
        long pageSize,
        @Schema(description = "total records")
        long totalRecords,
        @Schema(description = "total pages")
        long totalPages
) {
}